Added ESC = quit

+++ Labyrinth/src/Labyrinth/Events/KeyEvent.h
@@ -2,8 +2,6 @@
 
 #include "Event.h"
 
-#include "SDL.h"
-
 namespace Labyrinth {
 
 	class LABYRINTH_API KeyEvent : public Event
@@ -14,7 +12,7 @@ namespace Labyrinth {
 		EVENT_CLASS_CATEGORY(EventCategoryKeyboard | EventCategoryInput)
 
 	protected:
-		KeyEvent(int keycode) 
+		KeyEvent(int keycode)
 			: mKeyCode(keycode) {}
 
 		int mKeyCode;
@@ -23,7 +21,7 @@ namespace Labyrinth {
 	class LABYRINTH_API KeyPressedEvent : public KeyEvent
 	{
 	public:
-		KeyPressedEvent(Uint8 keycode, int repeatCount)
+		KeyPressedEvent(int keycode, int repeatCount)
 			: KeyEvent(keycode), mRepeatCount(repeatCount) {}
 
 		inline int GetRepeatCount() const { return mRepeatCount; }
@@ -43,7 +41,7 @@ namespace Labyrinth {
 	class LABYRINTH_API KeyReleasedEvent : public KeyEvent
 	{
 	public:
-		KeyReleasedEvent(Uint8 keycode)
+		KeyReleasedEvent(int keycode)
 			: KeyEvent(keycode) {}
 
 		std::string ToString() const override
